fix: Add error handling to post requests
Added back end logic for error handling pertaining to blog posts
and comments. There was logic for the front end, but not in
back end code.

Check if the request is coming from a user who is signed in
Check if the post exists
Check if the comment exists
Check if the signed in user is the author of the comment/post